# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Illuminae API Documentation",
        "description": "",
        "contact": {
            "email": "admin@admin.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://brain.test/",
            "description": "Local API Server"
        },
        {
            "url": "http://192.168.0.150:8008/",
            "description": "Local API Server Using IP"
        }
    ],
    "paths": {
        "/api/auth/login": {
            "post": {
                "tags": [
                    "auth,login"
                ],
                "summary": "authentication",
                "description": "",
                "operationId": "login",
                "parameters": [
                    {
                        "name": "username",
                        "in": "query",
                        "description": "Email Address",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/auth/forgetPasswordMail": {
            "post": {
                "tags": [
                    "auth,forgetPasswordMail"
                ],
                "summary": "authentication",
                "description": "",
                "operationId": "forgetPasswordMail",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/auth/verifyResetCode": {
            "post": {
                "tags": [
                    "auth,verifyResetCode"
                ],
                "summary": "authentication",
                "description": "",
                "operationId": "verifyResetCode",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "verification_code",
                        "in": "query",
                        "description": "Verification Code",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/auth/resetPassword": {
            "post": {
                "tags": [
                    "auth,resetPassword"
                ],
                "summary": "authentication",
                "description": "",
                "operationId": "resetPassword",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "verification_code",
                        "in": "query",
                        "description": "Verification Code",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password_confirmation",
                        "in": "query",
                        "description": "Password Confirmation",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}